requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.17.0
cryptography>=39.0.0
web3>=6.0.0
eth-account>=0.7.0
eth-utils>=2.1.0
merklelib>=1.0.0

# CLI tools
click>=8.1.0
typer>=0.7.0
rich>=13.0.0
tabulate>=0.9.0

# Container and orchestration
docker>=6.1.0
kubernetes>=26.1.0
docker-compose>=1.29.0

# Storage and database
pymongo>=4.3.0
redis>=4.5.0

# Network and API
fastapi>=0.95.0
uvicorn>=0.22.0
websockets>=11.0.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0
pytest-cov>=4.1.0
